First Commit
POP3 (Post Office Protocol version 3) is a standard protocol used for receiving email messages from a mail server. It is one of the most commonly used email protocols on the internet, and is supported by almost all email clients and mail servers.

POP3 is a client-server protocol, which means that the client (such as an email client application) connects to the server (the email server) to retrieve email messages. When a user logs into their email account using an email client, the client uses POP3 commands to connect to the server, authenticate the user's credentials, and retrieve the user's email messages.

POP3 works by downloading email messages from the server to the client, and then deleting them from the server. This means that once an email message has been downloaded to a client using POP3, it will no longer be available on the server. However, many email clients also offer the option to leave a copy of the email message on the server, or to delete messages from the server only after a certain period of time has elapsed.

POP3 uses TCP/IP for communication and typically uses port 110. However, some email servers also support SSL/TLS encryption for added security, in which case the port number used is typically 995.