Fourth Commit
HTTP stands for Hypertext Transfer Protocol, which is a protocol used for transmitting data over the internet. It is the foundation of data communication for the World Wide Web. HTTP defines how messages are formatted and transmitted, as well as how web servers and browsers should respond to various commands.

When a user types a URL (Uniform Resource Locator) into a web browser, the browser sends an HTTP request to the web server. The web server then responds with an HTTP status code, which indicates whether the request was successful or not. If the request is successful, the server sends back the requested data in the form of an HTTP response.

HTTP also supports various methods or verbs, including GET, POST, PUT, DELETE, HEAD, OPTIONS, and TRACE. These methods define the actions that can be performed on a web server.